moved get_config_hash to tools
