+++ flexget/utils/tools.py
@@ -1,5 +1,8 @@
 """Contains miscellaneous helpers"""
 from __future__ import unicode_literals, division, absolute_import
+
+from pprint import pformat
+
 from builtins import *  # pylint: disable=unused-import, redefined-builtin
 from future.moves.urllib import request
 from future.utils import PY2
@@ -525,3 +528,15 @@ def trim_dir(directory):
     file_name = os.path.join(directory, files[0])
     log.debug('removing least accessed file: %s', file_name)
     os.remove(file_name)
+
+
+def get_config_hash(config):
+    """
+    :param dict config: Configuration
+    :return: MD5 hash for *config*
+    """
+    if isinstance(config, dict) or isinstance(config, list):
+        # this does in fact support nested dicts, they're sorted too!
+        return hashlib.md5(pformat(config).encode('utf-8')).hexdigest()
+    else:
+        return hashlib.md5(str(config).encode('utf-8')).hexdigest()
